cid-api-django/requirements.txt
appdirs
django
djangorestframework
django_extensions
django-cors-headers
django-rest-registration
django-minio-storage>=0.2.1
django-spa
psycopg2-binary
packaging
pyparsing
pytz
Pillow
six
coreapi
